requirements.txt
bcrypt==3.1.7
DateTime==4.3
dnspython==1.16.0
env==0.1.0
Flask==1.1.1
Flask-Ext==0.1
Flask-Login==0.4.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
gym==0.15.4
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
json5==0.8.5
jsonpickle==1.2
jsonschema==3.0.2
lxml==4.4.1
pipenv==2018.11.26
pkginfo==1.5.0.1
plac==1.1.3
playsound==1.2.2
pluggy==0.13.0
ply==3.11
post==2019.4.13
preshed==3.0.2
request==2019.4.13
requests==2.22.0
requests-async==0.5.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
webexteamssdk==1.2
websocket-client==0.54.0
websockets==8.1
Werkzeug==0.16.0
WTForms==2.2.1
flask_paginate==0.6.0
python-dateutil==2.8.1